+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 ## Changelog
 
-### 0.1.1 (Next)
+### 0.1.1 (2018/03/27)
 
 * [#3](https://github.com/dblock/danger-toc/issues/3): Configure "Table of Contents" header with `toc.header` - [@dblock](https://github.com/dblock).
 * [#11](https://github.com/dblock/danger-toc/pull/11): Fix bug "undefined method #filename" in the method #toc_changes? - [@mr-dxdy](https://github.com/mr-dxdy).
-* Your contribution here.
 
 ### 0.1.0 (2017/11/5)